수정

+++ src/main/java/com/gang/domain/summoner/SummonerRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface SummonerRepository extends JpaRepository<SummonerEntity,Integer> {
     SummonerEntity findByName(String name);
-    SummonerEntity findBySummonerId(long summonerId);
+    SummonerEntity findBySummonerId(int summonerId);
 }
